DEV-726 models refactoring
