--- src/DirectorySync.Application/Models/GroupMapping.cs
@@ -1,8 +1,0 @@
-namespace DirectorySync.Application.Models
-{
-    public class GroupMapping
-    {
-        public string DirectoryGroup { get; set; } = string.Empty;
-        public string[] SignUpGroups { get; set; } = [];
-    }
-}
