+++ src/DirectorySync.Application/Models/Options/GroupMapping.cs
@@ -1,0 +1,7 @@
+namespace DirectorySync.Application.Models.Options;
+
+public class GroupMapping
+{
+    public string DirectoryGroup { get; set; } = string.Empty;
+    public string[] SignUpGroups { get; set; } = [];
+}
