--- src/DirectorySync.Application/Models/SyncSettings.cs
@@ -1,17 +1,0 @@
-namespace DirectorySync.Application.Models
-{
-    public class SyncSettings
-    {
-        public bool Enabled { get; set; }
-        public TimeSpan CloudConfigRefreshTimer { get; set; }
-        public TimeSpan SyncTimer { get; set; }
-        public TimeSpan ScanTimer { get; set; }
-
-        public TimeSpan TimeoutAd { get; set; }
-
-        public GroupMapping[] DirectoryGroupMappings { get; set; } = [];
-        public bool IncludeNestedGroups { get; set; } = true;
-
-        public PropsMapping PropertyMapping { get; set; } = new PropsMapping();
-    }
-}
